+++ src/sll/api/_generated.c
@@ -509,6 +509,66 @@ __SLL_EXTERNAL __SLL_CHECK_OUTPUT sll_object_t* sll_api_string_flip_case_raw(con
 
 
 
+__SLL_API_TYPE_sll_api_string_pad_left sll_api_string_pad_left(__SLL_API_ARGS_sll_api_string_pad_left);
+__SLL_EXTERNAL __SLL_CHECK_OUTPUT sll_object_t* sll_api_string_pad_left_raw(const sll_object_t*const* al,sll_arg_count_t all){
+	if (all<2){
+		return SLL_ACQUIRE_STATIC(str_zero);
+	}
+	const sll_object_t* a=*(al+0);
+	if (!(SLL_OBJECT_GET_TYPE(a)==SLL_OBJECT_TYPE_STRING)){
+		return SLL_ACQUIRE_STATIC(str_zero);
+	}
+	const sll_object_t* b=*(al+1);
+	if (!(SLL_OBJECT_GET_TYPE(b)==SLL_OBJECT_TYPE_INT)){
+		return SLL_ACQUIRE_STATIC(str_zero);
+	}
+	const sll_object_t* c=NULL;
+	if (all>2){
+		c=*(al+2);
+		if (!(SLL_OBJECT_GET_TYPE(c)==SLL_OBJECT_TYPE_CHAR)){
+			return SLL_ACQUIRE_STATIC(str_zero);
+		}
+	}
+	sll_string_t out;
+	sll_api_string_pad_left(&(a->dt.s),b->dt.i,(c?c->dt.c:SLL_NO_CHAR),&out);
+	sll_object_t* out_o=SLL_CREATE();
+	out_o->t=SLL_OBJECT_TYPE_STRING;
+	out_o->dt.s=out;
+	return out_o;
+}
+
+
+
+__SLL_API_TYPE_sll_api_string_pad_right sll_api_string_pad_right(__SLL_API_ARGS_sll_api_string_pad_right);
+__SLL_EXTERNAL __SLL_CHECK_OUTPUT sll_object_t* sll_api_string_pad_right_raw(const sll_object_t*const* al,sll_arg_count_t all){
+	if (all<2){
+		return SLL_ACQUIRE_STATIC(str_zero);
+	}
+	const sll_object_t* a=*(al+0);
+	if (!(SLL_OBJECT_GET_TYPE(a)==SLL_OBJECT_TYPE_STRING)){
+		return SLL_ACQUIRE_STATIC(str_zero);
+	}
+	const sll_object_t* b=*(al+1);
+	if (!(SLL_OBJECT_GET_TYPE(b)==SLL_OBJECT_TYPE_INT)){
+		return SLL_ACQUIRE_STATIC(str_zero);
+	}
+	const sll_object_t* c=NULL;
+	if (all>2){
+		c=*(al+2);
+		if (!(SLL_OBJECT_GET_TYPE(c)==SLL_OBJECT_TYPE_CHAR)){
+			return SLL_ACQUIRE_STATIC(str_zero);
+		}
+	}
+	sll_string_t out;
+	sll_api_string_pad_right(&(a->dt.s),b->dt.i,(c?c->dt.c:SLL_NO_CHAR),&out);
+	sll_object_t* out_o=SLL_CREATE();
+	out_o->t=SLL_OBJECT_TYPE_STRING;
+	out_o->dt.s=out;
+	return out_o;
+}
+
+
+
 __SLL_API_TYPE_sll_api_string_replace sll_api_string_replace(__SLL_API_ARGS_sll_api_string_replace);
 __SLL_EXTERNAL __SLL_CHECK_OUTPUT sll_object_t* sll_api_string_replace_raw(const sll_object_t*const* al,sll_arg_count_t all){
 	if (all<1){
@@ -786,6 +846,8 @@ static const internal_function_t _ifunc_data_ptr[]={
 	{"sll:string_convert",sll_api_string_convert_raw,SLL_INTERNAL_FUNCTION_FLAG_COMPILATION_CALL},
 	{"sll:string_format",sll_api_string_format_raw,SLL_INTERNAL_FUNCTION_FLAG_COMPILATION_CALL},
 	{"sll:string_flip_case",sll_api_string_flip_case_raw,SLL_INTERNAL_FUNCTION_FLAG_COMPILATION_CALL},
+	{"sll:string_pad_left",sll_api_string_pad_left_raw,SLL_INTERNAL_FUNCTION_FLAG_COMPILATION_CALL},
+	{"sll:string_pad_right",sll_api_string_pad_right_raw,SLL_INTERNAL_FUNCTION_FLAG_COMPILATION_CALL},
 	{"sll:string_replace",sll_api_string_replace_raw,SLL_INTERNAL_FUNCTION_FLAG_COMPILATION_CALL},
 	{"sll:string_reverse",sll_api_string_reverse_raw,SLL_INTERNAL_FUNCTION_FLAG_COMPILATION_CALL},
 	{"sll:string_split",sll_api_string_split_raw,SLL_INTERNAL_FUNCTION_FLAG_COMPILATION_CALL},
@@ -806,5 +868,5 @@ static const internal_function_t _ifunc_data_ptr[]={
 
 
 
-const sll_function_index_t _ifunc_size=43;
+const sll_function_index_t _ifunc_size=45;
 const internal_function_t* _ifunc_data=(const internal_function_t*)(&_ifunc_data_ptr);