+++ src/sll/api/string_api.c
@@ -332,6 +332,28 @@ __API_FUNC(string_flip_case){
 
 
 
+__API_FUNC(string_pad_left){
+	if (b<0){
+		sll_string_pad_right(a,(b<-((sll_integer_t)SLL_MAX_STRING_LENGTH)?SLL_MAX_STRING_LENGTH:(sll_string_length_t)(-b)),(c==SLL_NO_CHAR?' ':(sll_char_t)SLL_NO_CHAR),out);
+	}
+	else{
+		sll_string_pad_left(a,(b>SLL_MAX_STRING_LENGTH?SLL_MAX_STRING_LENGTH:(sll_string_length_t)b),(c==SLL_NO_CHAR?' ':(sll_char_t)SLL_NO_CHAR),out);
+	}
+}
+
+
+
+__API_FUNC(string_pad_right){
+	if (b<0){
+		sll_string_pad_right(a,(b<-((sll_integer_t)SLL_MAX_STRING_LENGTH)?SLL_MAX_STRING_LENGTH:(sll_string_length_t)(-b)),(c==SLL_NO_CHAR?' ':(sll_char_t)SLL_NO_CHAR),out);
+	}
+	else{
+		sll_string_pad_left(a,(b>SLL_MAX_STRING_LENGTH?SLL_MAX_STRING_LENGTH:(sll_string_length_t)b),(c==SLL_NO_CHAR?' ':(sll_char_t)SLL_NO_CHAR),out);
+	}
+}
+
+
+
 __API_FUNC(string_replace){
 	if (!b){
 		sll_string_clone(a,out);